:sparkles: remove unnecessary things round 2
