+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "windup-ui",
+  "name": "konveyor-static-report",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
