--- src/api/application.ts
@@ -1,10 +1,0 @@
-export interface ApplicationDto {
-  id: string;
-  name: string;
-  isVirtual: boolean;
-  tags: string[];
-  storyPoints: number;
-  incidents: {
-    [key: string]: number;
-  };
-}
