--- src/api/file-content.ts
@@ -1,4 +1,0 @@
-export interface FileContentDto {
-  id: string;
-  content: string;
-}
