--- src/api/label.ts
@@ -1,8 +1,0 @@
-export interface LabelDto {
-  id: string;
-  name: string;
-  description?: string;
-  supported: string[];
-  unsuitable: string[];
-  neutral: string[];
-}
