--- src/api/package-incidents.ts
@@ -1,4 +1,0 @@
-export interface ApplicationPackageIncidentsDto {
-  applicationId: string;
-  packages: { [key: string]: number };
-}
