--- src/api/rule-content.ts
@@ -1,4 +1,0 @@
-export interface RuleContentDto {
-  id: string;
-  content: string;
-}
