--- src/api/tag.ts
@@ -1,7 +1,0 @@
-export interface TagDto {
-  name: string;
-  title: string;
-  isRoot: boolean;
-  isPseudo: boolean;
-  parentsTagNames: string[];
-}
