+++ src/shared/components/index.ts
@@ -3,7 +3,7 @@ export * from "./simple-select";
 export * from "./simple-table-with-toolbar";
 // export * from "./issues-table";
 export * from "./dependencies-table";
-export * from "./rule-editor";
+// export * from "./rule-editor";
 export * from "./simple-markdown";
 export * from "./file-editor";
 export * from "./violations-table";
