--- src/shared/components/issues-table/index.ts
@@ -1,1 +1,0 @@
-export * from "./issues-table";
