--- src/shared/components/rule-editor/index.ts
@@ -1,1 +1,0 @@
-export * from "./rule-editor";
