--- src/utils/util.ts
@@ -1,7 +1,0 @@
-export const capitalizeFirstLetter = (val: string) => {
-  if (!val) {
-    return val;
-  }
-
-  return val.charAt(0).toUpperCase() + val.slice(1);
-};
